Caminadora_wlkata/requirements.txt
aioconsole==0.3.1
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
bleak==0.22.3
blessed==1.20.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmd2==1.5.0
colorama==0.4.6
coloredlogs==15.0.1
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
ph4-acmd2==1.0.0
ph4-walkingpad==1.0.2
psycopg2-binary==2.9.10
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreBluetooth==10.3.1
pyobjc-framework-libdispatch==10.3.1
pyperclip==1.9.0
pyserial==3.5
PyYAML==6.0.2
requests==2.32.3
setuptools==75.4.0
six==1.16.0
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
